+++ R/read_bed.R
@@ -7,7 +7,8 @@
 #'
 #' The code enforces several checks to validate data given the requested dimensions.
 #' Errors are thrown if file terminates too early or does not terminate after genotype matrix is filled.
-#' In addition, as each locus is encoded in an integer number of bytes, and each byte contains up to four individuals, bytes with fewer than four are padded with zeroes (non-zero pads throw errors).
+#' In addition, as each locus is encoded in an integer number of bytes, and each byte contains up to four individuals, bytes with fewer than four are padded.
+#' To agree with other software (plink2, BEDMatrix), byte padding values are ignored (may take on any value without causing errors).
 #'
 #' This function only supports locus-major BED files, which are the standard for modern data.
 #' Format is validated via the BED file's magic numbers (first three bytes of file).